HYPNO: refactored code, added basic debugging capabilities and fixes
